mod: 갤러리 접근 권한 파일명, 패키지 구조 수정
