+++ feature/profile/src/main/java/com/acon/acon/feature/profile/composable/screen/profileMod/composable/ProfileModScreen.kt
@@ -45,7 +45,7 @@ import com.acon.acon.core.designsystem.component.topbar.AconTopBar
 import com.acon.acon.core.designsystem.effect.LocalHazeState
 import com.acon.acon.core.designsystem.noRippleClickable
 import com.acon.acon.core.designsystem.theme.AconTheme
-import com.acon.acon.core.utils.feature.permission.storage.CheckAndRequestPhotoPermission
+import com.acon.acon.core.utils.feature.permission.media.CheckAndRequestMediaPermission
 import com.acon.acon.feature.profile.composable.component.GallerySelectBottomSheet
 import com.acon.acon.feature.profile.composable.component.NicknameValidMessageRow
 import com.acon.acon.feature.profile.composable.component.ProfilePhotoBox
@@ -110,7 +110,7 @@ internal fun ProfileModScreen(
             }
 
             if (state.requestPhotoPermission) {
-                CheckAndRequestPhotoPermission(
+                CheckAndRequestMediaPermission(
                     onPermissionGranted = {
                         onPhotoPermissionDenied()
                         onRequestProfileEditModal()
